+++ lib-debug/tracer.js
@@ -18,11 +18,17 @@ var _span_context2 = _interopRequireDefault(_span_context);
 
 var _constants = require('./constants');
 
-var _constants2 = _interopRequireDefault(_constants);
+var Constants = _interopRequireWildcard(_constants);
 
-var _reference = require('./reference');
+var _functions = require('./functions');
 
-var _reference2 = _interopRequireDefault(_reference);
+var Functions = _interopRequireWildcard(_functions);
+
+var _noop = require('./noop');
+
+var Noop = _interopRequireWildcard(_noop);
+
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -33,8 +39,17 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
  * implementation.
  *
  * The default object acts as a no-op implementation.
+ *
+ * Note to implementators: derived classes can choose to directly implement the
+ * methods in the "OpenTracing API methods" section, or optionally the subset of
+ * underscore-prefixed methods to pick up the argument checking and handling
+ * automatically from the base class.
  */
 var Tracer = function () {
+    function Tracer() {
+        _classCallCheck(this, Tracer);
+    }
+
     _createClass(Tracer, [{
         key: 'startSpan',
 
@@ -56,10 +71,7 @@ var Tracer = function () {
          *         reference: Tracer.childOf(parent.context()),
          *     });
          *
-         * @param {string|object} nameOrFields - if the given argument is a
-         *        string, it is the name of the operation and the second `fields`
-         *        argument is optional. If it is an object, it is treated as the
-         *        fields argument and a second argument should not be provided.
+         * @param {string} name - the name of the operation.
          * @param {object} [fields] - the fields to set on the newly created span.
          * @param {string} [fields.operationName] - the name to use for the newly
          *        created span. Required if called with a single argument.
@@ -80,95 +92,39 @@ var Tracer = function () {
          *        to represent time values with sub-millisecond accuracy.
          * @return {Span} - a new Span object.
          */
-        value: function startSpan(nameOrFields, fields) {
+        value: function startSpan(name, fields) {
+            // Debug-only runtime checks on the arguments
             if (process.env.NODE_ENV === 'debug') {
                 if (arguments.length > 2) {
                     throw new Error('Invalid number of arguments.');
                 }
-                if (typeof nameOrFields !== 'string' && (typeof nameOrFields === 'undefined' ? 'undefined' : _typeof(nameOrFields)) !== 'object') {
-                    throw new Error('argument expected to be a string or object');
+                if (typeof name !== 'string') {
+                    throw new Error('argument expected to be a string');
                 }
-                if (typeof nameOrFields === 'string' && nameOrFields.length === 0) {
+                if (name.length === 0) {
                     throw new Error('operation name cannot be length zero');
                 }
-                if ((typeof nameOrFields === 'undefined' ? 'undefined' : _typeof(nameOrFields)) === 'object') {
-                    if (arguments.length !== 1) {
-                        throw new Error('Unexpected number of arguments');
-                    }
-                    if (nameOrFields === null) {
-                        throw new Error('fields should not be null');
-                    }
-                    if (!nameOrFields.operationName) {
-                        throw new Error('operationName is a required parameter');
-                    }
+                if (fields && fields.childOf && fields.references) {
+                    throw new Error('At most one of `childOf` and ' + '`references` may be specified');
+                }
+                if (fields && fields.childOf && !(fields.childOf instanceof _span2.default || fields.childOf instanceof _span_context2.default)) {
+                    throw new Error('childOf must be a Span or SpanContext instance');
                 }
             }
 
-            var spanImp = null;
-            if (this._imp) {
-                // Normalize the argument so the implementation is always provided
-                // an associative array of fields.
-                if (arguments.length === 1) {
-                    if (typeof nameOrFields === 'string') {
-                        fields = {
-                            operationName: nameOrFields
-                        };
-                    } else {
-                        fields = nameOrFields;
-                    }
+            // Convert fields.childOf to fields.references as needed.
+            fields = fields || {};
+            if (fields.childOf) {
+                // Convert from a Span or a SpanContext into a Reference.
+                var childOf = Functions.childOf(fields.childOf);
+                if (fields.references) {
+                    fields.references.push(childOf);
                 } else {
-                    fields.operationName = nameOrFields;
+                    fields.references = [childOf];
                 }
-                if (process.env.NODE_ENV === 'debug') {
-                    if (fields.childOf && fields.references) {
-                        throw new Error('At most one of `childOf` and ' + '`references` may be specified');
-                    }
-                    if (fields.childOf && !(fields.childOf instanceof _span2.default || fields.childOf instanceof _span_context2.default)) {
-                        throw new Error('childOf must be a Span or SpanContext instance');
-                    }
-                }
-                // Convert fields.childOf to fields.references as needed.
-                if (fields.childOf) {
-                    // Convert from a Span or a SpanContext into a Reference.
-                    var childOf = this.childOf(fields.childOf);
-                    if (fields.references) {
-                        fields.references.push(childOf);
-                    } else {
-                        fields.references = [childOf];
-                    }
-                    delete fields.childOf;
-                }
-                spanImp = this._imp.startSpan(fields);
+                delete fields.childOf;
             }
-            return new _span2.default(spanImp);
-        }
-
-        /**
-         * Return a new REFERENCE_CHILD_OF reference.
-         *
-         * @param {SpanContext} spanContext - the parent SpanContext instance to
-         *        reference.
-         * @return a REFERENCE_CHILD_OF reference pointing to `spanContext`
-         */
-
-    }, {
-        key: 'childOf',
-        value: function childOf(spanContext) {
-            return new _reference2.default(_constants2.default.REFERENCE_CHILD_OF, spanContext);
-        }
-
-        /**
-         * Return a new REFERENCE_FOLLOWS_FROM reference.
-         *
-         * @param {SpanContext} spanContext - the parent SpanContext instance to
-         *        reference.
-         * @return a REFERENCE_FOLLOWS_FROM reference pointing to `spanContext`
-         */
-
-    }, {
-        key: 'followsFrom',
-        value: function followsFrom(spanContext) {
-            return new _reference2.default(_constants2.default.REFERENCE_FOLLOWS_FROM, spanContext);
+            return this._startSpan(name, fields);
         }
 
         /**
@@ -204,6 +160,7 @@ var Tracer = function () {
     }, {
         key: 'inject',
         value: function inject(spanContext, format, carrier) {
+            // Debug-only runtime checks on the arguments
             if (process.env.NODE_ENV === 'debug') {
                 if (arguments.length !== 3) {
                     throw new Error('Invalid number of arguments.');
@@ -214,24 +171,22 @@ var Tracer = function () {
                 if (typeof format !== 'string') {
                     throw new Error('format expected to be a string. Found: ' + (typeof format === 'undefined' ? 'undefined' : _typeof(format)));
                 }
-                if (format === _constants2.default.FORMAT_TEXT_MAP && (typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) !== 'object') {
+                if (format === Constants.FORMAT_TEXT_MAP && (typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) !== 'object') {
                     throw new Error('Unexpected carrier object for FORMAT_TEXT_MAP');
                 }
-                if (format === _constants2.default.FORMAT_HTTP_HEADERS && (typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) !== 'object') {
+                if (format === Constants.FORMAT_HTTP_HEADERS && (typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) !== 'object') {
                     throw new Error('Unexpected carrier object for FORMAT_HTTP_HEADERS');
                 }
-                if (format === _constants2.default.FORMAT_BINARY && (typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) !== 'object') {
+                if (format === Constants.FORMAT_BINARY && (typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) !== 'object') {
                     throw new Error('Unexpected carrier object for FORMAT_BINARY');
                 }
             }
 
-            if (this._imp) {
-                // Allow the user to pass a Span instead of a SpanContext
-                if (spanContext instanceof _span2.default) {
-                    spanContext = spanContext.context();
-                }
-                this._imp.inject(spanContext._imp, format, carrier);
+            // Allow the user to pass a Span instead of a SpanContext
+            if (spanContext instanceof _span2.default) {
+                spanContext = spanContext.context();
             }
+            return this._inject(spanContext, format, carrier);
         }
 
         /**
@@ -260,6 +215,7 @@ var Tracer = function () {
     }, {
         key: 'extract',
         value: function extract(format, carrier) {
+            // Debug-only runtime checks on the arguments
             if (process.env.NODE_ENV === 'debug') {
                 if (arguments.length !== 2) {
                     throw new Error('Invalid number of arguments.');
@@ -267,88 +223,50 @@ var Tracer = function () {
                 if (typeof format !== 'string' || !format.length) {
                     throw new Error('format is expected to be a string of non-zero length');
                 }
-                if (format === _constants2.default.FORMAT_TEXT_MAP && !((typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) === 'object')) {
+                if (format === Constants.FORMAT_TEXT_MAP && !((typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) === 'object')) {
                     throw new Error('Unexpected carrier object for FORMAT_TEXT_MAP');
                 }
-                if (format === _constants2.default.FORMAT_HTTP_HEADERS && !((typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) === 'object')) {
+                if (format === Constants.FORMAT_HTTP_HEADERS && !((typeof carrier === 'undefined' ? 'undefined' : _typeof(carrier)) === 'object')) {
                     throw new Error('Unexpected carrier object for FORMAT_HTTP_HEADERS');
                 }
-                if (format === _constants2.default.FORMAT_BINARY) {
+                if (format === Constants.FORMAT_BINARY) {
                     if (carrier.buffer !== undefined && _typeof(carrier.buffer) !== 'object') {
                         throw new Error('Unexpected carrier object for FORMAT_BINARY');
                     }
                 }
             }
-            var spanContextImp = null;
-            if (this._imp) {
-                spanContextImp = this._imp.extract(format, carrier);
-            }
-            if (spanContextImp !== null) {
-                return new _span_context2.default(spanContextImp);
-            }
-            return null;
+            return this._extract(format, carrier);
         }
 
-        /**
-         * Request that any buffered or in-memory data is flushed out of the process.
-         *
-         * @param {function(err: objectg)} done - optional callback function with
-         *        the signature `function(err)` that will be called as soon as the
-         *        flush completes. `err` should be null or undefined if the flush
-         *        was successful.
-         */
+        // ---------------------------------------------------------------------- //
+        // Derived classes can choose to implement the below
+        // ---------------------------------------------------------------------- //
+
+        // NOTE: the input to this method is *always* an associative array. The
+        // public-facing startSpan() method normalizes the arguments so that
+        // all N implementations do not need to worry about variations in the call
+        // signature.
+        //
+        // The default behavior returns a no-op span.
 
     }, {
-        key: 'flush',
-        value: function flush(done) {
-            if (process.env.NODE_ENV === 'debug') {
-                if (arguments.length > 1) {
-                    throw new Error('Invalid number of arguments');
-                }
-                if (done !== undefined && typeof done !== 'function') {
-                    throw new Error('callback expected to be a function');
-                }
-            }
-            if (!this._imp) {
-                done(null);
-                return;
-            }
-            this._imp.flush(done);
+        key: '_startSpan',
+        value: function _startSpan(name, fields) {
+            return Noop.span;
         }
 
-        // ---------------------------------------------------------------------- //
-        // Private and non-standard methods
-        // ---------------------------------------------------------------------- //
+        // The default behavior is a no-op.
 
-        /**
-         * Note: this constructor should not be called directly by consumers of this
-         * code. The singleton's initNewTracer() method should be invoked instead.
-         */
+    }, {
+        key: '_inject',
+        value: function _inject(spanContext, format, carrier) {}
 
-    }]);
+        // The default behavior is to return null.
 
-    function Tracer(imp) {
-        _classCallCheck(this, Tracer);
-
-        this._imp = imp || null;
-    }
-
-    /**
-     * Handle to implementation object.
-     *
-     * Use of this method is discouraged as it greatly reduces the portability of
-     * the calling code. Use only when implementation-specific functionality must
-     * be used and cannot accessed otherwise.
-     *
-     * @return {object}
-     *         An implementation-dependent object.
-     */
-
-
-    _createClass(Tracer, [{
-        key: 'imp',
-        value: function imp() {
-            return this._imp;
+    }, {
+        key: '_extract',
+        value: function _extract(format, carrier) {
+            return Noop.spanContext;
         }
     }]);