+++ src/span_context.js
@@ -1,5 +1,3 @@
-'use strict';
-
 /**
  * SpanContext represents Span state that must propagate to descendant Spans
  * and across process boundaries.
@@ -11,21 +9,5 @@
  * <trace_id, span_id, sampled> tuple).
  */
 export default class SpanContext {
-    /**
-     * Constructs a new SpanContext object.
-     *
-     * This method should not be called directly; Span.context() should be used
-     * instead.
-     */
-    constructor(imp) {
-        this._imp = imp;
-    }
-
-    /**
-     * Returns the SpanContext implementation object. The returned object is by
-     * its nature entirely implementation-dependent.
-     */
-    imp() {
-        return this._imp;
-    }
+    // The SpanContext is entirely implementation dependent
 }