+++ src/tracer.js
@@ -1,15 +1,19 @@
-'use strict';
-
 import Span from './span';
 import SpanContext from './span_context';
-import Constants from './constants';
-import Reference from './reference';
+import * as Constants from './constants';
+import * as Functions from './functions';
+import * as Noop from './noop';
 
 /**
  * Tracer is the entry-point between the instrumentation API and the tracing
  * implementation.
  *
  * The default object acts as a no-op implementation.
+ *
+ * Note to implementators: derived classes can choose to directly implement the
+ * methods in the "OpenTracing API methods" section, or optionally the subset of
+ * underscore-prefixed methods to pick up the argument checking and handling
+ * automatically from the base class.
  */
 export default class Tracer {
 
@@ -30,10 +34,7 @@ export default class Tracer {
      *         reference: Tracer.childOf(parent.context()),
      *     });
      *
-     * @param {string|object} nameOrFields - if the given argument is a
-     *        string, it is the name of the operation and the second `fields`
-     *        argument is optional. If it is an object, it is treated as the
-     *        fields argument and a second argument should not be provided.
+     * @param {string} name - the name of the operation.
      * @param {object} [fields] - the fields to set on the newly created span.
      * @param {string} [fields.operationName] - the name to use for the newly
      *        created span. Required if called with a single argument.
@@ -54,92 +55,42 @@ export default class Tracer {
      *        to represent time values with sub-millisecond accuracy.
      * @return {Span} - a new Span object.
      */
-    startSpan(nameOrFields, fields) {
+    startSpan(name, fields) {
+        // Debug-only runtime checks on the arguments
         if (process.env.NODE_ENV === 'debug') {
             if (arguments.length > 2) {
                 throw new Error('Invalid number of arguments.');
             }
-            if (typeof nameOrFields !== 'string' && typeof nameOrFields !== 'object') {
-                throw new Error('argument expected to be a string or object');
+            if (typeof name !== 'string') {
+                throw new Error('argument expected to be a string');
             }
-            if (typeof nameOrFields === 'string' && nameOrFields.length === 0) {
+            if (name.length === 0) {
                 throw new Error('operation name cannot be length zero');
             }
-            if (typeof nameOrFields === 'object') {
-                if (arguments.length !== 1) {
-                    throw new Error('Unexpected number of arguments');
-                }
-                if (nameOrFields === null) {
-                    throw new Error('fields should not be null');
-                }
-                if (!nameOrFields.operationName) {
-                    throw new Error('operationName is a required parameter');
-                }
+            if (fields && fields.childOf && fields.references) {
+                throw new Error('At most one of `childOf` and ' +
+                        '`references` may be specified');
+            }
+            if (fields && fields.childOf && !(
+                        fields.childOf instanceof Span ||
+                        fields.childOf instanceof SpanContext)) {
+                throw new Error('childOf must be a Span or SpanContext instance');
             }
         }
 
-        let spanImp = null;
-        if (this._imp) {
-            // Normalize the argument so the implementation is always provided
-            // an associative array of fields.
-            if (arguments.length === 1) {
-                if (typeof nameOrFields === 'string') {
-                    fields = {
-                        operationName : nameOrFields,
-                    };
-                } else {
-                    fields = nameOrFields;
-                }
+        // Convert fields.childOf to fields.references as needed.
+        fields = fields || {};
+        if (fields.childOf) {
+            // Convert from a Span or a SpanContext into a Reference.
+            let childOf = Functions.childOf(fields.childOf);
+            if (fields.references) {
+                fields.references.push(childOf);
             } else {
-                fields.operationName = nameOrFields;
+                fields.references = [childOf];
             }
-            if (process.env.NODE_ENV === 'debug') {
-                if (fields.childOf && fields.references) {
-                    throw new Error('At most one of `childOf` and ' +
-                            '`references` may be specified');
-                }
-                if (fields.childOf && !(
-                            fields.childOf instanceof Span ||
-                            fields.childOf instanceof SpanContext)) {
-                    throw new Error('childOf must be a Span or SpanContext instance');
-                }
-            }
-            // Convert fields.childOf to fields.references as needed.
-            if (fields.childOf) {
-                // Convert from a Span or a SpanContext into a Reference.
-                let childOf = this.childOf(fields.childOf);
-                if (fields.references) {
-                    fields.references.push(childOf);
-                } else {
-                    fields.references = [childOf];
-                }
-                delete(fields.childOf);
-            }
-            spanImp = this._imp.startSpan(fields);
+            delete(fields.childOf);
         }
-        return new Span(spanImp);
-    }
-
-    /**
-     * Return a new REFERENCE_CHILD_OF reference.
-     *
-     * @param {SpanContext} spanContext - the parent SpanContext instance to
-     *        reference.
-     * @return a REFERENCE_CHILD_OF reference pointing to `spanContext`
-     */
-    childOf(spanContext) {
-        return new Reference(Constants.REFERENCE_CHILD_OF, spanContext);
-    }
-
-    /**
-     * Return a new REFERENCE_FOLLOWS_FROM reference.
-     *
-     * @param {SpanContext} spanContext - the parent SpanContext instance to
-     *        reference.
-     * @return a REFERENCE_FOLLOWS_FROM reference pointing to `spanContext`
-     */
-    followsFrom(spanContext) {
-        return new Reference(Constants.REFERENCE_FOLLOWS_FROM, spanContext);
+        return this._startSpan(name, fields);
     }
 
     /**
@@ -172,6 +123,7 @@ export default class Tracer {
      *         for a description of the carrier object.
      */
     inject(spanContext, format, carrier) {
+        // Debug-only runtime checks on the arguments
         if (process.env.NODE_ENV === 'debug') {
             if (arguments.length !== 3) {
                 throw new Error('Invalid number of arguments.');
@@ -193,13 +145,11 @@ export default class Tracer {
             }
         }
 
-        if (this._imp) {
-            // Allow the user to pass a Span instead of a SpanContext
-            if (spanContext instanceof Span) {
-                spanContext = spanContext.context();
-            }
-            this._imp.inject(spanContext._imp, format, carrier);
+        // Allow the user to pass a Span instead of a SpanContext
+        if (spanContext instanceof Span) {
+            spanContext = spanContext.context();
         }
+        return this._inject(spanContext, format, carrier);
     }
 
     /**
@@ -225,6 +175,7 @@ export default class Tracer {
      *         be found in `carrier`
      */
     extract(format, carrier) {
+        // Debug-only runtime checks on the arguments
         if (process.env.NODE_ENV === 'debug') {
             if (arguments.length !== 2) {
                 throw new Error('Invalid number of arguments.');
@@ -244,64 +195,29 @@ export default class Tracer {
                 }
             }
         }
-        let spanContextImp = null;
-        if (this._imp) {
-            spanContextImp = this._imp.extract(format, carrier);
-        }
-        if (spanContextImp !== null) {
-            return new SpanContext(spanContextImp);
-        }
-        return null;
+        return this._extract(format, carrier);
     }
-
-    /**
-     * Request that any buffered or in-memory data is flushed out of the process.
-     *
-     * @param {function(err: objectg)} done - optional callback function with
-     *        the signature `function(err)` that will be called as soon as the
-     *        flush completes. `err` should be null or undefined if the flush
-     *        was successful.
-     */
-    flush(done) {
-        if (process.env.NODE_ENV === 'debug') {
-            if (arguments.length > 1) {
-                throw new Error('Invalid number of arguments');
-            }
-            if (done !== undefined && typeof done !== 'function') {
-                throw new Error('callback expected to be a function');
-            }
-        }
-        if (!this._imp) {
-            done(null);
-            return;
-        }
-        this._imp.flush(done);
-    }
-
 
     // ---------------------------------------------------------------------- //
-    // Private and non-standard methods
+    // Derived classes can choose to implement the below
     // ---------------------------------------------------------------------- //
 
-    /**
-     * Note: this constructor should not be called directly by consumers of this
-     * code. The singleton's initNewTracer() method should be invoked instead.
-     */
-    constructor(imp) {
-        this._imp = imp || null;
+    // NOTE: the input to this method is *always* an associative array. The
+    // public-facing startSpan() method normalizes the arguments so that
+    // all N implementations do not need to worry about variations in the call
+    // signature.
+    //
+    // The default behavior returns a no-op span.
+    _startSpan(name, fields) {
+        return Noop.span;
     }
 
-    /**
-     * Handle to implementation object.
-     *
-     * Use of this method is discouraged as it greatly reduces the portability of
-     * the calling code. Use only when implementation-specific functionality must
-     * be used and cannot accessed otherwise.
-     *
-     * @return {object}
-     *         An implementation-dependent object.
-     */
-    imp() {
-        return this._imp;
+    // The default behavior is a no-op.
+    _inject(spanContext, format, carrier) {
+    }
+
+    // The default behavior is to return null.
+    _extract(format, carrier) {
+        return Noop.spanContext;
     }
 }